+++ tests/integration/test_signed_tt.py
@@ -129,7 +129,7 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert span_server.attributes["TriggeredTrace"] == True
         assert "this-will-be-ignored" not in span_server.attributes
 
-        # Check root span tracestate has `sw` and `xtrace_options_response` keys
+        # Check client span tracestate has `sw` and `xtrace_options_response` keys
         # In this test we know `sw` value will also have invalid span_id.
         # SWO APM uses TraceState to stash the trigger trace response so it's available 
         # at the time of custom injecting the x-trace-options-response header.
@@ -168,8 +168,8 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
            and start of the trace.
         2. Some traceparent and tracestate are injected into service's outgoing request
            (done by OTel TraceContextTextMapPropagator).
-        3. No x-trace-options-response header is calculated because the extracted
-           x-trace-options does not include trigger-trace.
+        3. x-trace-options-response header is calculated because there is an extracted
+           x-trace-options header
         4. Sampling-related, SWKeys, and custom-*, and attributes are set
            for the root/service entry span, but not what's ignored nor TriggeredTrace.
         5. The span_id of the outgoing request span matches the span_id portion in the
@@ -225,11 +225,9 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert new_trace_flags == "01"
 
         assert "tracestate" in resp_json
-        # In this test we know there is `sw` in tracestate
-        # where value will be new_span_id and new_trace_flags.
-        # There should be no `xtrace_options_response` key because there is
-        # no trigger-trace in the extracted x-trace-options header.
-        assert resp_json["tracestate"] == "sw={}-{}".format(
+        # In this test we know there is `sw` and `xtrace_options_response` in
+        # tracestate where value of former will be new_span_id and new_trace_flags.
+        assert resp_json["tracestate"] == "sw={}-{},xtrace_options_response=auth####ok;trigger-trace####not-requested;ignored####this-will-be-ignored".format(
             new_span_id,
             new_trace_flags,
         )
@@ -250,12 +248,11 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert span_client.name == "HTTP GET"
         assert span_client.kind == trace_api.SpanKind.CLIENT
 
-        # Check root span tracestate has `sw` key.
+        # Check root span tracestate has `sw` and `xtrace_options_response` keys
         # In this test we know `sw` value will have invalid span_id.
-        # There should be no `xtrace_options_response` key because there is
-        # no trigger-trace in the extracted x-trace-options header.
         expected_trace_state = trace_api.TraceState([
             ("sw", "0000000000000000-01"),
+            ("xtrace_options_response", "auth####ok;trigger-trace####not-requested;ignored####this-will-be-ignored"),
         ])
         assert span_server.context.trace_state == expected_trace_state
 
@@ -281,12 +278,11 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert "TriggeredTrace" not in span_server.attributes
         assert "this-will-be-ignored" not in span_server.attributes
 
-        # Check root span tracestate has `sw` key/
-        # In this test we know `sw` value will also have invalid span_id.
-        # There should be no `xtrace_options_response` key because there is
-        # no trigger-trace in the extracted x-trace-options header.
+        # Check client span tracestate has `sw` and `xtrace_options_response` keys
+        # In this test we know `sw` value will have invalid span_id.
         expected_trace_state = trace_api.TraceState([
             ("sw", "0000000000000000-01"),
+            ("xtrace_options_response", "auth####ok;trigger-trace####not-requested;ignored####this-will-be-ignored"),
         ])
         assert span_client.context.trace_state == expected_trace_state